move ClientRequestContext to wrapper
